006_POC_Main_Agent : Moved packages to be more consistent
