+++ src/main/java/io/github/stealingdapenta/idletd/custommob/CustomMobHandler.java
@@ -1,4 +1,4 @@
-package io.github.stealingdapenta.idletd.service.custommob;
+package io.github.stealingdapenta.idletd.custommob;
 
 import io.github.stealingdapenta.idletd.Idletd;
 import lombok.RequiredArgsConstructor;
